+++ CHANGELOG.md
@@ -1,6 +1,12 @@
 Change log
 ==========
 
+1.2.9
+-----
+
+* 🎁 Set default aiport to flightplan _destination_
+* 🎁 Add link to world clock
+
 1.2.8
 -----